requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
altair>=5.2
gspread>=6.1
google-auth>=2.35
google-api-python-client>=2.151
python-dateutil>=2.9
reportlab>=4.2
vl-convert-python>=1.7.0
altair-saver>=0.5.0
